api/requirements.txt
flask
flask_cors
mysql-connector-python
python-dotenv
sqlalchemy
groq
requests